SCI32: Fix palette color overflow

+++ engines/sci/graphics/palette32.h
@@ -265,10 +265,15 @@ private:
 	 * The fade table records the expected intensity level of each pixel
 	 * in the palette that will be displayed on the next frame.
 	 */
-	byte _fadeTable[256];
+	uint16 _fadeTable[256];
 
 public:
-	void setFade(const uint8 percent, const uint8 fromColor, const uint16 toColor);
+	/**
+	 * Sets the intensity level for a range of palette
+	 * entries. An intensity of zero indicates total
+	 * darkness. Intensity may be set to over 100 percent.
+	 */
+	void setFade(const uint16 percent, const uint8 fromColor, const uint16 toColor);
 	void fadeOff();
 	void applyFade();
 };
